+++ src/Enums/Provider.php
@@ -7,4 +7,13 @@ enum Provider: string
     case GOOGLE = 'google';
     case FACEBOOK = 'facebook';
     case SEZNAM = 'seznam';
+
+    public static function enabledCases(): array
+    {
+        return array_values(
+            array_filter(self::cases(), fn(self $provider): bool =>
+                config('service.' . $provider->value . '.enabled', false)
+            )
+        );
+    }
 }